% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{database_fields}
\alias{database_fields}
\title{Collection of arguments to "@inheritParams" from}
\arguments{
\item{conn}{Database connection.}

\item{calstate_id}{Integer vector of \code{calibration_state} ID.}

\item{calstate_datetime}{POSIXct vector of date and time of calibration.}

\item{calstate_parameter}{String vector of values of calibration parameters.}

\item{calstate_comment}{String vector of additional comments.}

\item{dev_id}{Integer vector of \code{device} ID.}

\item{dev_name}{String vector of name of device.}

\item{dev_identifier}{String vector of device identifiers, e.g. serial
numbers.}

\item{dev_comment}{String vector of additional comments.}

\item{devman_id}{Integer vector of \code{device_manufacturer} IDs.}

\item{devman_name}{String vector of name of device manufacturer.}

\item{devman_comment}{String vector of additional comments.}

\item{devmod_id}{Integer vector of \code{device_model} ID.}

\item{devmod_name}{String vector of name of model.}

\item{devmod_comment}{String vector of additional comments.}

\item{devtype_id}{Integer vector of \code{device_type} IDs.}

\item{devtype_name}{String vector of name of device type.}

\item{devtype_comment}{String vector of additional comments.}

\item{int_id}{Integer vector of \code{integration} ID.}

\item{int_measurement_interval}{Numeric vector of intervals between
measurements in s.}

\item{int_interval}{Numeric vector of integration interval in s of one stored
measurement.}

\item{int_comment}{Character vector of additional information.}

\item{inttype_id}{Integer vector of \code{integration_type} ID.}

\item{inttype_name}{String vector of name of integration type.}

\item{inttype_description}{String vector of description of
\code{integration_type}.}

\item{inttype_comment}{String vector of additional comments.}

\item{md_id}{Integer vector of measurand ID.}

\item{md_comment}{Character vector of additional information.}

\item{md_name}{Character vector of measurand name.}

\item{md_setup_datetime}{POSIXct vector of date and time of set-up of
measurand.}

\item{md_height}{Numeric vector of measurement height.}

\item{md_orientation}{Numeric vector of north-south measurement orientation.}

\item{md_tilt}{Numeric vector of measurement tilt.}

\item{name}{Name of the table.}

\item{pers_id}{Integer vector of \code{person} ID.}

\item{pers_name}{Character vector of person name.}

\item{pers_comment}{Character vector of additional information.}

\item{standard_name}{Standard name of a physical quantity as defined by the
CF convention.}

\item{pq_id}{Integer vector of \code{physical_quantity} ID.}

\item{pq_name}{String vector of name of physical quantity.}

\item{pq_unit}{String vector of units of physical quantity. Use "1" for
unitless quantities.}

\item{pq_description}{String vector of description of physical quantity.}

\item{pq_comment}{String vector of additional comments.}

\item{qf_id}{Integer vector with 1 <= qf_id <= 9 indicating value ok and
qf_id >= 10 indicating value not ok.}

\item{qf_name}{Character vector of quality_flag name.}

\item{qf_description}{Character vector of quality_flag description.}

\item{qf_comment}{Character vector of of additional information.}

\item{site_id}{Integer vector of \code{site} ID.}

\item{site_name}{String vector of name of site or name of campaign.}

\item{site_lat}{Numeric vector of geographical latitude WGS84.}

\item{site_lon}{Numeric vector of geographical longitude WGS84.}

\item{site_altitude}{Numeric vector of height above sea level of surface in
m.}

\item{site_comment}{String vector of additional comments.}

\item{stadl_id}{Integer vector of \code{station_adlershof} ID.}

\item{stadl_datetime}{POSIXct vector of date and time of measurement.}

\item{stadl_value}{Numeric vector of measurement values.}

\item{stadlcor_datetime}{POSIXct vector of corrected date and time of
measurement.}

\item{stadlcor_value}{Numeric vector of corrected value of measurement.}

\item{stapa_id}{Integer vector of \code{station_patagonia} ID.}

\item{stapa_datetime}{POSIXct vector of date and time of measurement.}

\item{stapa_value}{Numeric vector of measurement values.}

\item{stapacor_datetime}{POSIXct vector of corrected date and time of
measurement.}

\item{stapacor_value}{Numeric vector of corrected value of measurement.}
}
\description{
Collection of arguments to "@inheritParams" from
}
\keyword{internal}
