% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{dbAdd_calibrated_device}
\alias{dbAdd_calibrated_device}
\alias{dbWriteTable_calibrated_device}
\title{Add calibrated devices to \code{calibrated_device} table}
\usage{
dbAdd_calibrated_device(conn, dev_name, caldev_datetime = NULL,
  caldev_parameter = NULL, caldev_comment = NULL)

dbWriteTable_calibrated_device(conn, dev_id, caldev_datetime = NULL,
  caldev_parameter = NULL, caldev_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{dev_name}{String vector of name of device.}

\item{caldev_datetime}{POSIXct vector of date and time of calibration.}

\item{caldev_parameter}{String vector of values of calibration parameters.}

\item{caldev_comment}{String vector of additional comments.}

\item{dev_id}{Integer vector of \code{device} ID.}
}
\value{
Data frame of newly added rows.
}
\description{
\code{dbWriteTable_calibrated_device} requires a correct device id while
\code{dbAdd_calibrated_device} derives that from the device name.
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_device_manufacturer(con, "TSI")
dbWriteTable_device_type(con, "thermometer")
dbWriteTable_device_model(con, "THERMO1000", 1, 1)
dbWriteTable_device(con, "My first THERMO1000", 1, "NCC1701-T")
dbWriteTable_calibrated_device(con,
                               1,
                               as.POSIXct("2012-01-01 12:15:12", tz = "GMT"),
                               "a=10, b=99.12")
dbDisconnect(con)
}
}
\seealso{
\code{\link{dbAdd_uncalibrated_device}} for adding a device that
  does not require calibration to both \code{device} and
  \code{calibrated_device}

Other custom dbWriteTable functions: \code{\link{dbAddCorrection_station_adlershof}},
  \code{\link{dbAdd_device_model}},
  \code{\link{dbAdd_device}},
  \code{\link{dbAdd_integration}},
  \code{\link{dbAdd_measurand}},
  \code{\link{dbAdd_physical_quantity}},
  \code{\link{dbAdd_station_adlershof}},
  \code{\link{dbWriteTable_device_manufacturer}},
  \code{\link{dbWriteTable_device_type}},
  \code{\link{dbWriteTable_integration_type}},
  \code{\link{dbWriteTable_person}},
  \code{\link{dbWriteTable_quality_flag}},
  \code{\link{dbWriteTable_site}}
}
