% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{dbAdd_device_model}
\alias{dbAdd_device_model}
\alias{dbWriteTable_device_model}
\title{Add device models to \code{device_model} table}
\usage{
dbAdd_device_model(conn, devmod_name, devtype_name, devman_name = NULL,
  devmod_comment = NULL)

dbWriteTable_device_model(conn, devmod_name, devtype_id, devman_id = NULL,
  devmod_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{devmod_name}{String vector of name of model.}

\item{devtype_name}{String vector of name of device type.}

\item{devman_name}{String vector of name of device manufacturer.}

\item{devmod_comment}{String vector of additional comments.}

\item{devtype_id}{Integer vector of \code{device_type} IDs.}

\item{devman_id}{Integer vector of \code{device_manufacturer} IDs.}
}
\value{
Data frame of newly added rows.
}
\description{
\code{dbWriteTable_device_model} requires a correct device type id and,
optionally, a correct device manufacturer id while \code{dbAdd_device_model}
derives that from the device type and the device manufacturer name,
respectively.
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_device_manufacturer(con, "TSI")
dbWriteTable_device_type(con, "thermometer")
dbWriteTable_device_model(con, "THERMO1000", 1, 1)
dbDisconnect(con)
}
}
\seealso{
Other custom dbWriteTable functions: \code{\link{dbAddCorrection_station_adlershof}},
  \code{\link{dbAdd_calibrated_device}},
  \code{\link{dbAdd_device}},
  \code{\link{dbAdd_integration}},
  \code{\link{dbAdd_measurand}},
  \code{\link{dbAdd_physical_quantity}},
  \code{\link{dbAdd_station_adlershof}},
  \code{\link{dbWriteTable_device_manufacturer}},
  \code{\link{dbWriteTable_device_type}},
  \code{\link{dbWriteTable_integration_type}},
  \code{\link{dbWriteTable_person}},
  \code{\link{dbWriteTable_quality_flag}},
  \code{\link{dbWriteTable_site}}
}
