% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{dbAdd_device_uncalibrated}
\alias{dbAdd_device_uncalibrated}
\title{Insert data into \code{device} and \code{calibration_state} without
calibration parameters}
\usage{
dbAdd_device_uncalibrated(
  conn,
  dev_name,
  devmod_name,
  dev_identifier = NULL,
  dev_comment = NULL
)
}
\arguments{
\item{conn}{Database connection.}

\item{dev_name}{String vector of name of device.}

\item{devmod_name}{String vector of name of model.}

\item{dev_identifier}{String vector of device identifiers, e.g. serial
numbers.}

\item{dev_comment}{String vector of additional comments.}
}
\value{
List of data frames of newly added rows.
}
\description{
This function adds new devices to the database that do not require any
calibration. First, it adds the new devices into the \code{device} table. It
then uses the respective created \code{device.dev_id} to add new entries in
\code{calibration_state} with both \code{calibration_state.calstate_datetime}
and \code{calibration_state.calstate_parameter} being \code{NA}.
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_device_manufacturer(con, "TSI")
dbWriteTable_device_type(con, "thermometer")
dbWriteTable_device_model(con, "THERMO1000", 1, 1)
dbAdd_device_uncalibrated(con, "My first THERMO1000", 1, "NCC1701-T")
dbDisconnect(con)
}
}
\seealso{
\code{\link{dbWriteTable_device}} and
  \code{\link{dbWriteTable_calibration_state}} for the separate functions
  called by this wrapper.
}
