% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{dbAdd_measurand}
\alias{dbAdd_measurand}
\alias{dbWriteTable_measurand}
\title{Add measurand to \code{measurand} table}
\usage{
dbAdd_measurand(conn, md_name, md_setup_datetime, pq_name, site_name, dev_name,
  int_id, md_height = NULL, md_orientation = NULL, md_tilt = NULL,
  pers_name = NULL, md_comment = NULL)

dbWriteTable_measurand(conn, md_name, md_setup_datetime, pq_id, site_id,
  calstate_id, int_id, md_height = NULL, md_orientation = NULL,
  md_tilt = NULL, pers_id = NULL, md_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{md_name}{Character vector of measurand name.}

\item{md_setup_datetime}{POSIXct vector of date and time of set-up of
measurand.}

\item{pq_name}{String vector of name of physical quantity.}

\item{site_name}{String vector of name of site or name of campaign.}

\item{dev_name}{String vector of name of device.}

\item{int_id}{Integer vector of \code{integration} ID.}

\item{md_height}{Numeric vector of measurement height.}

\item{md_orientation}{Numeric vector of north-south measurement orientation.}

\item{md_tilt}{Numeric vector of measurement tilt.}

\item{pers_name}{Character vector of person name.}

\item{md_comment}{Character vector of additional information.}

\item{pq_id}{Integer vector of \code{physical_quantity} ID.}

\item{site_id}{Integer vector of \code{site} ID.}

\item{calstate_id}{Integer vector of \code{calibration_state} ID.}

\item{pers_id}{Integer vector of \code{person} ID.}
}
\value{
Data frame of newly added rows.
}
\description{
\code{dbWriteTable_measurand} requires a correct physical quantity id, site
id, calibrated device id, integration id and, optionally, person id.
\code{dbAdd_measurand} derives the physical quantity id, site id, calibrated
device id and person id from the respective names; integration id still have
to be entered numerical as id. \strong{Note} that \code{dbAdd_measurand}
derives the calibrated device id from the device id with the most recent
calibration date \code{calstate_datetime}. Use
\code{\link{dbReadTable_integration_detail}} to query the integration table
and select the correct id.
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
# add all required entries before with the respective dbWriteTable_*
dbWriteTable_measurand(md_name = "TA2M_1",
                       md_setup_datetime = as.POSIXct("2012-01-01 12:15:12", tz = "UTC"),
                       pq_id = 1,
                       site_id = 1,
                       calstate_id = 1,
                       int_id = 1,
                       md_height = 2.,
                       pers_id = 1,
                       md_comment = "the 2m temperature"))
dbDisconnect(con)
}
}
\seealso{
Other custom dbWriteTable functions: \code{\link{dbAddCorrection_station_adlershof}},
  \code{\link{dbAdd_calibration_state}},
  \code{\link{dbAdd_device_model}},
  \code{\link{dbAdd_device}},
  \code{\link{dbAdd_integration}},
  \code{\link{dbAdd_physical_quantity}},
  \code{\link{dbAdd_station_adlershof}},
  \code{\link{dbWriteTable_device_manufacturer}},
  \code{\link{dbWriteTable_device_type}},
  \code{\link{dbWriteTable_integration_type}},
  \code{\link{dbWriteTable_person}},
  \code{\link{dbWriteTable_quality_flag}},
  \code{\link{dbWriteTable_site}}
}
