% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{dbAdd_physical_quantity}
\alias{dbAdd_physical_quantity}
\alias{dbWriteTable_physical_quantity}
\title{Add a physical quantity into \code{physical_quantity} table}
\usage{
dbAdd_physical_quantity(conn, pq_name, pq_comment = NULL)

dbWriteTable_physical_quantity(conn, pq_name, pq_unit,
  pq_description = NULL, pq_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{pq_name}{String vector of name of physical quantity.}

\item{pq_comment}{String vector of additional comments.}

\item{pq_unit}{String vector of units of physical quantity. Use "1" for
unitless quantities.}

\item{pq_description}{String vector of description of physical quantity.}
}
\value{
Data frame of newly added rows.
}
\description{
These function add rows defining a physical quantity into
\code{physical_quantity}. \strong{Follow the
\href{http://cfconventions.org/Data/cf-standard-names/48/build/cf-standard-name-table.html}{CF
conventions} as far as possible.} \code{dbAdd_physical_quantity} supports
this by getting both the unit and the description from valid standard names
set with \code{pq_name}. \code{dbWriteTable_physical_quantity} allows
entering the strings manually, however, it does not enforce valid CF strings.
}
\details{
\code{dbAdd_physical_quantity} queries
\url{http://cfconventions.org/Data/cf-standard-names/48/src/cf-standard-name-table.xml}
once per session. Thus, it requires an internet connection.
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_physical_quantity(con, "air temperature", "degC")
dbDisconnect(con)
}
}
\seealso{
Other custom dbWriteTable functions: \code{\link{dbAddCorrection_station_adlershof}},
  \code{\link{dbAddCorrection_station_patagonia}},
  \code{\link{dbAdd_calibration_state}},
  \code{\link{dbAdd_device_model}},
  \code{\link{dbAdd_device}},
  \code{\link{dbAdd_integration}},
  \code{\link{dbAdd_measurand}},
  \code{\link{dbAdd_station_adlershof}},
  \code{\link{dbAdd_station_patagonia}},
  \code{\link{dbWriteTable_device_manufacturer}},
  \code{\link{dbWriteTable_device_type}},
  \code{\link{dbWriteTable_integration_type}},
  \code{\link{dbWriteTable_person}},
  \code{\link{dbWriteTable_quality_flag}},
  \code{\link{dbWriteTable_site}}
}
\concept{custom dbWriteTable functions}
