% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{dbAdd_station_patagonia}
\alias{dbAdd_station_patagonia}
\alias{dbWriteTable_station_patagonia}
\title{Add measurements to \code{station_patagonia} table}
\usage{
dbAdd_station_patagonia(conn, md_name, stapa_datetime, stapa_value,
  qf_id = NULL)

dbWriteTable_station_patagonia(conn, stapa_datetime, md_id, stapa_value,
  qf_id = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{md_name}{Character vector of measurand name.}

\item{stapa_datetime}{POSIXct vector of date and time of measurement.}

\item{stapa_value}{Numeric vector of measurement values.}

\item{qf_id}{Integer vector with 1 <= qf_id <= 9 indicating value ok and
qf_id >= 10 indicating value not ok.}

\item{md_id}{Integer vector of measurand ID.}
}
\value{
For performance reason, contrary to the meta data table functions,
  this function does not return anything.
}
\description{
\code{dbWriteTable_station_patagonia} requires a correct measurand id while
\code{dbAdd_station_patagonia} derives that from the measurand name.
}
\details{
\code{dbWriteTable_station_patagonia} requires a correct measurand id
\code{md_id} to write measurements to station_patagonia.
\code{dbAdd_station_patagonia} finds \code{md_id} from a given measurand name
\code{md_name}. To this end, it queries the table \code{measurand} and
selects the according to \code{md_setup_datetime} the most recent
\code{md_id}.
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
# add all required entries before with the respective dbWriteTable_*
dbAdd_station_patagonia(con,
                        md_name = "TA2M_1"
                        stapa_datetime = as.POSIXct("2017-01-01 12:05:12", tz = "UTC"),
                        stapa_value = 272.15)
dbWriteTable_station_patagonia(con,
                               stapa_datetime = as.POSIXct("2017-01-01 12:15:12", tz = "UTC"),
                               md_id = 1,
                               stapa_value = 273.15)
dbDisconnect(con)
}
}
\seealso{
Other custom dbWriteTable functions: \code{\link{dbAddCorrection_station_adlershof}},
  \code{\link{dbAddCorrection_station_patagonia}},
  \code{\link{dbAdd_calibration_state}},
  \code{\link{dbAdd_device_model}},
  \code{\link{dbAdd_device}},
  \code{\link{dbAdd_integration}},
  \code{\link{dbAdd_measurand}},
  \code{\link{dbAdd_physical_quantity}},
  \code{\link{dbAdd_station_adlershof}},
  \code{\link{dbWriteTable_device_manufacturer}},
  \code{\link{dbWriteTable_device_type}},
  \code{\link{dbWriteTable_integration_type}},
  \code{\link{dbWriteTable_person}},
  \code{\link{dbWriteTable_quality_flag}},
  \code{\link{dbWriteTable_site}}
}
\concept{custom dbWriteTable functions}
