% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dbAdd_uncalibrated_device}
\alias{dbAdd_uncalibrated_device}
\title{Insert data into \code{device} and \code{calibrated_device} without
calibration parameters}
\usage{
dbAdd_uncalibrated_device(conn, dev_name, devmod_id, dev_identifier = NULL,
  dev_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{dev_name}{String vector of name of device.}

\item{devmod_id}{Integer vector of \code{device_model} ID.}

\item{dev_identifier}{String vector of device identifiers, e.g. serial
numbers.}

\item{dev_comment}{String vector of additional comments.}
}
\description{
This function adds new devices to the database that do not require any
calibration. First, it adds the new devices into the \code{device} table. It
then uses the respective created \code{device.dev_id} to add new entries in
\code{calibrated_device} with both \code{calibrated_device.caldev_datetime}
and \code{calibrated_device.caldev_parameter} being \code{NULL}.
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_device_manufacturer(con, "TSI")
dbWriteTable_device_type(con, "thermometer")
dbWriteTable_device_model(con, "THERMO1000", 1, 1)
dbAdd_uncalibrated_device(con, "My first THERMO1000", 1, "NCC1701-T")
dbDisconnect(con)
}
}
