% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connection.R
\name{dbConnect_klimageo}
\alias{dbConnect_klimageo}
\title{Connect to klimageo database}
\usage{
dbConnect_klimageo(dsn = "klimageodb")
}
\arguments{
\item{dsn}{The Data Source Name. This corresponds to the set-up entry in the
\href{http://db.rstudio.com/best-practices/drivers/}{odbc.ini}.}
}
\value{
An S4 object that inherits from DBIConnection. This object is used to
  communicate with the database engine.
}
\description{
Connect to the klimegeo database. The connection is done via the \code{odbc}
package. Thus, the odbc driver and connection settings have to be set-up
outside of R before. The correct timezone is chosen by this command.
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbDisconnect(con)
}
}
