% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{dbBegin_or_Savepoint}
\alias{dbBegin_or_Savepoint}
\alias{dbCommit_or_Savepoint}
\alias{dbRollback_or_Savepoint}
\alias{dbWithTransaction_or_Savepoint}
\title{Self-contained SQL commands within new transactions or savepoint in present
transaction}
\usage{
dbBegin_or_Savepoint(conn, spname)

dbCommit_or_Savepoint(conn, transaction, spname)

dbRollback_or_Savepoint(conn, transaction, spname)

dbWithTransaction_or_Savepoint(conn, code, spname)
}
\arguments{
\item{conn}{Database connection.}

\item{spname}{String of the possibly required savpoint.}

\item{transaction}{Use standard transactions or use Savepoints?}

\item{code}{An arbitrary block of R code.}
}
\value{
\code{dbWithTransaction()} returns the value of the executed code.
  Failure to initiate the transaction (e.g. if the connection is closed)
  gives an error.
}
\description{
The function \code{dbWithTransaction_or_Savepoint()} starts a new transaction
or issues a savepoint via \code{dbBegin_or_Savepoint()}. If the execution of
the \code{code} argument fails, the state will be committed or the savepoint
removed via \code{dbCommit_or_Savepoint()}. Otherwise, the transaction is
rollbacked completely or up until the savepoint via
\code{dbRollback_or_Savepoint()}. The functions work similiarly to
\code{\link[DBI]{dbWithTransaction}()}, \code{\link[DBI]{dbBegin}()},
\code{\link[DBI]{dbCommit}()} or \code{\link[DBI]{dbRollback}()},
respectively, but allow to be nested by using savepoints if necessary.
}
\keyword{internal}
