% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dbWriteTable_calibrated_device}
\alias{dbWriteTable_calibrated_device}
\title{Insert data into \code{calibrated_device} table}
\usage{
dbWriteTable_calibrated_device(conn, dev_id, caldev_datetime = NULL,
  caldev_parameter = NULL, caldev_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{dev_id}{Integer vector of device ID.}

\item{caldev_datetime}{POSIXct vector of date and time of calibration.}

\item{caldev_parameter}{String vector of values of calibration parameters.'}

\item{caldev_comment}{String vector of additional comments.}
}
\description{
Insert data into \code{calibrated_device} table
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_device_manufacturer(con, "TSI")
dbWriteTable_device_type(con, "thermometer")
dbWriteTable_device_model(con, "THERMO1000", 1, 1)
dbWriteTable_device(con, "My first THERMO1000", 1, "NCC1701-T")
dbWriteTable_calibrated_device(con,
                               1,
                               as.POSIXct("2012-01-01 12:15:12", tz = "GMT"),
                               "a=10, b=99.12")
dbDisconnect(conn)
}
}
