% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{dbWriteTable_device_manufacturer}
\alias{dbWriteTable_device_manufacturer}
\title{Insert data into \code{device_manufacturer} table}
\usage{
dbWriteTable_device_manufacturer(conn, devman_name,
  devman_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{devman_name}{String vector of name of device manufacturer.}

\item{devman_comment}{String vector of additional comments.}
}
\value{
Data frame of newly added rows.
}
\description{
Insert data into \code{device_manufacturer} table
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_device_manufacturer(con, "TSI")
dbDisconnect(con)
}
}
\seealso{
Other custom dbWriteTable functions: \code{\link{dbAddCorrection_station_adlershof}},
  \code{\link{dbAddCorrection_station_patagonia}},
  \code{\link{dbAdd_calibration_state}},
  \code{\link{dbAdd_device_model}},
  \code{\link{dbAdd_device}},
  \code{\link{dbAdd_integration}},
  \code{\link{dbAdd_measurand}},
  \code{\link{dbAdd_physical_quantity}},
  \code{\link{dbAdd_station_adlershof}},
  \code{\link{dbAdd_station_patagonia}},
  \code{\link{dbWriteTable_device_type}},
  \code{\link{dbWriteTable_integration_type}},
  \code{\link{dbWriteTable_person}},
  \code{\link{dbWriteTable_quality_flag}},
  \code{\link{dbWriteTable_site}}
}
\concept{custom dbWriteTable functions}
