% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{dbWriteTable_device_model}
\alias{dbWriteTable_device_model}
\title{Insert data into \code{device_model} table}
\usage{
dbWriteTable_device_model(conn, devmod_name, devtype_id, devman_id = NULL,
  devmod_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{devmod_name}{String vector of name of model.}

\item{devtype_id}{Integer vector of \code{device_type} IDs.}

\item{devman_id}{Integer vector of \code{device_manufacturer} IDs.}

\item{devmod_comment}{String vector of additional comments.}
}
\value{
Data frame of newly added rows.
}
\description{
Insert data into \code{device_model} table
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_device_manufacturer(con, "TSI")
dbWriteTable_device_type(con, "thermometer")
dbWriteTable_device_model(con, "THERMO1000", 1, 1)
dbDisconnect(con)
}
}
\seealso{
Other custom dbWriteTable functions: \code{\link{dbAddCorrection_station_adlershof}},
  \code{\link{dbAddMeasurement_station_adlershof}},
  \code{\link{dbWriteTable_calibrated_device}},
  \code{\link{dbWriteTable_device_manufacturer}},
  \code{\link{dbWriteTable_device_type}},
  \code{\link{dbWriteTable_device}},
  \code{\link{dbWriteTable_integration_type}},
  \code{\link{dbWriteTable_integration}},
  \code{\link{dbWriteTable_measurand}},
  \code{\link{dbWriteTable_person}},
  \code{\link{dbWriteTable_physical_quantity}},
  \code{\link{dbWriteTable_quality_flag}},
  \code{\link{dbWriteTable_site}}
}
