% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dbWriteTable_device_type}
\alias{dbWriteTable_device_type}
\title{Insert data into \code{device_type} table}
\usage{
dbWriteTable_device_type(conn, devtype_name, devtype_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{devtype_name}{String vector of name of device type.}

\item{devtype_comment}{String vector of additional comments.}
}
\description{
Insert data into \code{device_type} table
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_device_type(con, "thermometer")
dbDisconnect(con)
}
}
