% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dbWriteTable_integration}
\alias{dbWriteTable_integration}
\title{Insert data into \code{integration} table}
\usage{
dbWriteTable_integration(conn, inttype_id, int_measurement_interval,
  int_interval, int_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{inttype_id}{Integer vector of \code{integration_type} ID.}

\item{int_measurement_interval}{Numeric vector of intervals between measurements in s.}

\item{int_interval}{Numeric vector of integration interval in s of one stored measurement.}

\item{int_comment}{Character vector of additional information.}
}
\description{
Insert data into \code{integration} table
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_integration_type("average", "average value over a period")
dbWriteTable_integration(con, 1, 60, 600)
dbDisconnect(con)
}
}
