% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dbWriteTable_integration_type}
\alias{dbWriteTable_integration_type}
\title{Insert data into \code{integration_type} table}
\usage{
dbWriteTable_integration_type(conn, inttype_name, inttype_description,
  inttype_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{inttype_name}{String vector of name of integration type.}

\item{inttype_description}{String vector of description of \code{integration_type}.}

\item{inttype_comment}{String vector of additional comments.}
}
\description{
Insert data into \code{integration_type} table
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_integration_type(con, "average", "average value over a period")
dbDisconnect(con)
}
}
