% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{dbWriteTable_measurand}
\alias{dbWriteTable_measurand}
\title{Insert data into \code{measurand} table}
\usage{
dbWriteTable_measurand(conn, md_name, md_setup_datetime, pq_id, site_id,
  caldev_id, int_id, md_height = NULL, pers_id = NULL, md_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{md_name}{Character vector of measurand name.}

\item{md_setup_datetime}{POSIXct vector of date and time of set-up of measurand.}

\item{pq_id}{Integer vector of \code{physical_quantity} ID.}

\item{site_id}{Integer vector of \code{site} ID.}

\item{caldev_id}{Integer vector of \code{calibrated_device} ID.}

\item{int_id}{Integer vector of \code{integration} ID.}

\item{md_height}{Numeric vector of measurement height.}

\item{pers_id}{Integer vector of \code{person} ID.}

\item{md_comment}{Character vector of additional information.}
}
\value{
Data frame of newly added rows.
}
\description{
Insert data into \code{measurand} table
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
# add all required entries before with the respective dbWriteTable_*
dbWriteTable_measurand(md_name = "TA2M_1",
                       md_setup_datetime = as.POSIXct("2012-01-01 12:15:12", tz = "UTC"),
                       pq_id = 1,
                       site_id = 1,
                       caldev_id = 1,
                       int_id = 1,
                       md_height = 2.,
                       pers_id = 1,
                       md_comment = "the 2m temperature"))
dbDisconnect(con)
}
}
\seealso{
Other custom dbWriteTable functions: \code{\link{dbWriteTable_calibrated_device}},
  \code{\link{dbWriteTable_device_manufacturer}},
  \code{\link{dbWriteTable_device_model}},
  \code{\link{dbWriteTable_device_type}},
  \code{\link{dbWriteTable_device}},
  \code{\link{dbWriteTable_integration_type}},
  \code{\link{dbWriteTable_integration}},
  \code{\link{dbWriteTable_person}},
  \code{\link{dbWriteTable_physical_quantity}},
  \code{\link{dbWriteTable_quality_flag}},
  \code{\link{dbWriteTable_site}},
  \code{\link{dbWriteTable_station_adlershof}}
}
