% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dbWriteTable_person}
\alias{dbWriteTable_person}
\title{Insert data into \code{person} table}
\usage{
dbWriteTable_person(conn, pers_name, pers_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{pers_name}{Character vector of person name.}

\item{pers_comment}{Character vector of additional information.}
}
\description{
Insert data into \code{person} table
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_person(con, "Maxi Mueller")
dbDisconnect(con)
}
}
