% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{dbWriteTable_physical_quantity}
\alias{dbWriteTable_physical_quantity}
\title{Insert data into \code{physical_quantity} table}
\usage{
dbWriteTable_physical_quantity(conn, pq_name, pq_unit, pq_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{pq_name}{String vector of name of physical quantity.}

\item{pq_unit}{String vector of units of physical quantity. Use "1" for
unitless quantities.}

\item{pq_comment}{String vector of additional comments.}
}
\value{
Data frame of newly added rows.
}
\description{
Insert data into \code{physical_quantity} table
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_physical_quantity(con, "air temperature", "degC")
dbDisconnect(con)
}
}
\seealso{
Other custom dbWriteTable functions: \code{\link{dbAddCorrection_station_adlershof}},
  \code{\link{dbAdd_station_adlershof}},
  \code{\link{dbWriteTable_calibrated_device}},
  \code{\link{dbWriteTable_device_manufacturer}},
  \code{\link{dbWriteTable_device_model}},
  \code{\link{dbWriteTable_device_type}},
  \code{\link{dbWriteTable_device}},
  \code{\link{dbWriteTable_integration_type}},
  \code{\link{dbWriteTable_integration}},
  \code{\link{dbWriteTable_measurand}},
  \code{\link{dbWriteTable_person}},
  \code{\link{dbWriteTable_quality_flag}},
  \code{\link{dbWriteTable_site}}
}
