% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dbWriteTable_site}
\alias{dbWriteTable_site}
\title{Insert data into \code{site} table}
\usage{
dbWriteTable_site(conn, site_name, site_lat = NULL, site_lon = NULL,
  site_altitude = NULL, site_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{site_name}{String vector of name of site or name of campaign.}

\item{site_lat}{Numeric vector of geographical latitude WGS84.}

\item{site_lon}{Numeric vector of geographical longitude WGS84.}

\item{site_altitude}{Numeric vector of height above sea level of surface in
m.}

\item{site_comment}{String vector of additional comments.}
}
\description{
Insert data into \code{site} table
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_site(con, "Adlershof", site_lon = 14., site_lat = 53.)
dbDisconnect(con)
}
}
