% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dbWriteTable_station_adlershof}
\alias{dbWriteTable_station_adlershof}
\title{Insert data into \code{station_adlershof} table}
\usage{
dbWriteTable_station_adlershof(conn, stadl_datetime, md_id, stadl_value,
  qf_id = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{stadl_datetime}{POSIXct vector of date and time of measurement.}

\item{md_id}{Integer vector of measurand ID.}

\item{stadl_value}{Numeric vector of measurement values.}

\item{qf_id}{Integer vector with 1 <= qf_id <= 9 indicating value ok and
qf_id >= 10 indicating value not ok;}
}
\description{
Insert data into \code{station_adlershof} table
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
# add all required entries before with the respective dbWriteTable_*
dbWriteTable_station_adlershof
dbDisconnect(con)
}
}
