% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_data.R
\name{get_cf_unit_description}
\alias{get_cf_unit_description}
\title{Get unit and description of a variable from CF standard name table}
\usage{
get_cf_unit_description(standard_name, version = 48)
}
\arguments{
\item{standard_name}{String vector of standard names in the CF standard name
table.}

\item{version}{The version of the CF standard name table.}
}
\value{
A data frame with columns \code{unit} and \code{description}.
}
\description{
This function extracts for a \code{standard_name} the unit and description
form the CF standard name table.
}
\examples{
get_cf_unit_description("air_temperature")
# use older standard
get_cf_unit_description("air_temperature", version = 42)
}
\keyword{internal}
