% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_column_values}
\alias{get_column_values}
\title{Get values of a column from a table}
\usage{
get_column_values(conn, name, column)
}
\arguments{
\item{conn}{Database connection.}

\item{name}{Name of the table.}

\item{column}{Name of the column.}
}
\value{
Vector of values.
}
\description{
Get values of a column from a table
}
