% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{get_columns}
\alias{get_columns}
\title{Get columns as a data frame from a table}
\usage{
get_columns(conn, name, columns)
}
\arguments{
\item{conn}{Database connection.}

\item{name}{Name of the table.}

\item{columns}{Character vector of names of the column.}
}
\value{
Data frame of columns.
}
\description{
Get columns as a data frame from a table
}
\keyword{internal}
