% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_primarykey}
\alias{get_primarykey}
\title{Get primary key name of a table}
\usage{
get_primarykey(conn, name)
}
\arguments{
\item{conn}{Database connection.}

\item{name}{Name of the table.}
}
\value{
String of primary key.
}
\description{
Get primary key name of a table
}
\keyword{internal}
