% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{readtable}
\alias{readtable}
\alias{dbReadTable_site}
\alias{dbReadTable_device_manufacturer}
\alias{dbReadTable_device_type}
\alias{dbReadTable_device_model}
\alias{dbReadTable_device}
\alias{dbReadTable_calibrated_device}
\alias{dbReadTable_physical_quantity}
\alias{dbReadTable_integration_type}
\alias{dbReadTable_integration}
\alias{dbReadTable_person}
\alias{dbReadTable_measurand}
\alias{dbReadTable_quality_flag}
\alias{dbReadTable_station_adlershof}
\alias{dbReadTable_station_adlershof_correction}
\title{Table reading}
\usage{
dbReadTable_site(conn)

dbReadTable_device_manufacturer(conn)

dbReadTable_device_type(conn)

dbReadTable_device_model(conn)

dbReadTable_device(conn)

dbReadTable_calibrated_device(conn)

dbReadTable_physical_quantity(conn)

dbReadTable_integration_type(conn)

dbReadTable_integration(conn)

dbReadTable_person(conn)

dbReadTable_measurand(conn)

dbReadTable_quality_flag(conn)

dbReadTable_station_adlershof(conn)

dbReadTable_station_adlershof_correction(conn)
}
\arguments{
\item{conn}{Database connection.}
}
\description{
These function retrieve the respective tables as a data frame from the
database.
}
\seealso{
\code{\link[DBI]{dbReadTable}} for the general function to retrieve
  complete tables. \code{\link[DBI]{dbGetQuery}} and
  \url{http://db.rstudio.com/best-practices/run-queries-safely/} for more
  general SQL queries. Alternatively, you can use \code{\link[dplyr]{tbl}} of
  the \code{dplyr} package to only work on part of the database.
}
