% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{dbWriteTable_integration}
\alias{dbWriteTable_integration}
\alias{dbAdd_integration}
\title{Add integration kind to \code{integration} table}
\usage{
dbWriteTable_integration(conn, inttype_id, int_measurement_interval,
  int_interval, int_comment = NULL)

dbAdd_integration(conn, inttype_name, int_measurement_interval, int_interval,
  int_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{inttype_id}{Integer vector of \code{integration_type} ID.}

\item{int_measurement_interval}{Numeric vector of intervals between
measurements in s.}

\item{int_interval}{Numeric vector of integration interval in s of one stored
measurement.}

\item{int_comment}{Character vector of additional information.}

\item{inttype_name}{String vector of name of integration type.}
}
\value{
Data frame of newly added rows.
}
\description{
\code{dbWriteTable_integration} requires a correct device integration type id
while \code{dbAdd_integration} derives that from the device model name.
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_integration_type("average", "average value over a period")
dbWriteTable_integration(con, 1, 60, 600)
dbDisconnect(con)
}
}
\seealso{
Other custom dbWriteTable functions: \code{\link{dbAddCorrection_station_adlershof}},
  \code{\link{dbAdd_station_adlershof}},
  \code{\link{dbWriteTable_calibrated_device}},
  \code{\link{dbWriteTable_device_manufacturer}},
  \code{\link{dbWriteTable_device_model}},
  \code{\link{dbWriteTable_device_type}},
  \code{\link{dbWriteTable_device}},
  \code{\link{dbWriteTable_integration_type}},
  \code{\link{dbWriteTable_measurand}},
  \code{\link{dbWriteTable_person}},
  \code{\link{dbWriteTable_physical_quantity}},
  \code{\link{dbWriteTable_quality_flag}},
  \code{\link{dbWriteTable_site}}
}
