% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dbWriteTable_device}
\alias{dbWriteTable_device}
\title{Insert data into \code{device} table}
\usage{
dbWriteTable_device(conn, dev_name, devmod_id, dev_identifier = NULL,
  dev_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{dev_name}{String vector of name of device.}

\item{devmod_id}{Integer vector of \code{device_model} ID.}

\item{dev_identifier}{String vector of device identifiers, e.g. serial
numbers.}

\item{dev_comment}{String vector of additional comments.}
}
\description{
Insert data into \code{device} table
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_device_manufacturer(con, "TSI")
dbWriteTable_device_type(con, "thermometer")
dbWriteTable_device_model(con, "THERMO1000", 1, 1)
dbWriteTable_device(con, "My first THERMO1000", 1, "NCC1701-T")
dbDisconnect(con)
}
}
