% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dbWriteTable_device_manufacturer}
\alias{dbWriteTable_device_manufacturer}
\title{Insert data into \code{device_manufacturer} table}
\usage{
dbWriteTable_device_manufacturer(conn, devman_name, devman_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{devman_name}{String vector of name of device manufacturer.}

\item{devman_comment}{String vector of additional comments.}
}
\description{
Insert data into \code{device_manufacturer} table
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_device_manufacturer(con, "TSI")
dbDisconnect(conn)
}
}
