% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dbWriteTable_device_model}
\alias{dbWriteTable_device_model}
\title{Insert data into \code{device_model} table}
\usage{
dbWriteTable_device_model(conn, devmod_name, devtype_id, devman_id = NULL,
  devmod_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{devmod_name}{String vector of name of model.}

\item{devtype_id}{Integer vector of IDs from \code{device_type}.}

\item{devman_id}{Integer vector of IDs from \code{device_manufacturer}.}

\item{devmod_comment}{String vector of additional comments.}
}
\description{
Insert data into \code{device_model} table
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_device_manufacturer(con, "TSI")
dbWriteTable_device_type(con, "thermometer")
dbWriteTable_device_model(con, "THERMO1000", 1, 1)
dbDisconnect(conn)
}
}
