% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{dbWriteTable_person}
\alias{dbWriteTable_person}
\title{Insert data into \code{person} table}
\usage{
dbWriteTable_person(conn, pers_name, pers_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{pers_name}{Character vector of person name.}

\item{pers_comment}{Character vector of additional information.}
}
\value{
Data frame of newly added rows.
}
\description{
Insert data into \code{person} table
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_person(con, "Maxi Mueller")
dbDisconnect(con)
}
}
\seealso{
Other custom dbWriteTable functions: \code{\link{dbAddCorrection_station_adlershof}},
  \code{\link{dbAdd_station_adlershof}},
  \code{\link{dbWriteTable_calibrated_device}},
  \code{\link{dbWriteTable_device_manufacturer}},
  \code{\link{dbWriteTable_device_model}},
  \code{\link{dbWriteTable_device_type}},
  \code{\link{dbWriteTable_device}},
  \code{\link{dbWriteTable_integration_type}},
  \code{\link{dbWriteTable_integration}},
  \code{\link{dbWriteTable_measurand}},
  \code{\link{dbWriteTable_physical_quantity}},
  \code{\link{dbWriteTable_quality_flag}},
  \code{\link{dbWriteTable_site}}
}
