% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dbWriteTable_physical_quantity}
\alias{dbWriteTable_physical_quantity}
\title{Insert data into \code{physical_quantity} table}
\usage{
dbWriteTable_physical_quantity(conn, pq_name, pq_unit, pq_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{pq_name}{String vector of name of physical quantity.}

\item{pq_unit}{String vector of units of physical quantity. Use "1" for
unitless quantities.}

\item{pq_comment}{String vector of additional comments.}
}
\description{
Insert data into \code{physical_quantity} table
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_physical_quantity(con, "air temperature", "degC")
dbDisconnect(con)
}
}
