% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{dbWriteTable_site}
\alias{dbWriteTable_site}
\title{Insert data into \code{site} table}
\usage{
dbWriteTable_site(conn, site_name, site_lat = NULL, site_lon = NULL,
  site_altitude = NULL, site_comment = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{site_name}{String vector of name of site or name of campaign.}

\item{site_lat}{Numeric vector of geographical latitude WGS84.}

\item{site_lon}{Numeric vector of geographical longitude WGS84.}

\item{site_altitude}{Numeric vector of height above sea level of surface in
m.}

\item{site_comment}{String vector of additional comments.}
}
\value{
Data frame of newly added rows.
}
\description{
Insert data into \code{site} table
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
dbWriteTable_site(con, "Adlershof", site_lon = 14., site_lat = 53.)
dbDisconnect(con)
}
}
\seealso{
Other custom dbWriteTable functions: \code{\link{dbAddCorrection_station_adlershof}},
  \code{\link{dbAdd_calibrated_device}},
  \code{\link{dbAdd_device_model}},
  \code{\link{dbAdd_device}},
  \code{\link{dbAdd_integration}},
  \code{\link{dbAdd_measurand}},
  \code{\link{dbAdd_physical_quantity}},
  \code{\link{dbAdd_station_adlershof}},
  \code{\link{dbWriteTable_device_manufacturer}},
  \code{\link{dbWriteTable_device_type}},
  \code{\link{dbWriteTable_integration_type}},
  \code{\link{dbWriteTable_person}},
  \code{\link{dbWriteTable_quality_flag}}
}
