% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write.R
\name{dbWriteTable_station_adlershof}
\alias{dbWriteTable_station_adlershof}
\title{Insert data into \code{station_adlershof} table}
\usage{
dbWriteTable_station_adlershof(conn, stadl_datetime, md_id, stadl_value,
  qf_id = NULL)
}
\arguments{
\item{conn}{Database connection.}

\item{stadl_datetime}{POSIXct vector of date and time of measurement.}

\item{md_id}{Integer vector of measurand ID.}

\item{stadl_value}{Numeric vector of measurement values.}

\item{qf_id}{Integer vector with 1 <= qf_id <= 9 indicating value ok and
qf_id >= 10 indicating value not ok.}
}
\value{
For performance reason, contrary to the meta data table functions,
  this function does not return anything.
}
\description{
Insert data into \code{station_adlershof} table
}
\examples{
\dontrun{
con <- dbConnect_klimageo()
# add all required entries before with the respective dbWriteTable_*
dbWriteTable_station_adlershof(con,
                               stadl_datetime = as.POSIXct("2017-01-01 12:15:12", tz = "UTC"),
                               md_id = 1,
                               stadl_value = 273.15)
dbDisconnect(con)
}
}
\seealso{
Other custom dbWriteTable functions: \code{\link{dbWriteTable_calibrated_device}},
  \code{\link{dbWriteTable_device_manufacturer}},
  \code{\link{dbWriteTable_device_model}},
  \code{\link{dbWriteTable_device_type}},
  \code{\link{dbWriteTable_device}},
  \code{\link{dbWriteTable_integration_type}},
  \code{\link{dbWriteTable_integration}},
  \code{\link{dbWriteTable_measurand}},
  \code{\link{dbWriteTable_person}},
  \code{\link{dbWriteTable_physical_quantity}},
  \code{\link{dbWriteTable_quality_flag}},
  \code{\link{dbWriteTable_site}}
}
